@@ -11,12 +11,12 @@
   "author": "",
   "license": "AGPL-3.0-only",
   "devDependencies": {
-    "chai": "^5.2.0",
-    "eslint": "^9.22.0",
-    "mocha": "^11.1.0",
-    "selenium-webdriver": "^4.29.0"
+    "chai": "^6.2.0",
+    "eslint": "^9.37.0",
+    "mocha": "^11.7.4",
+    "selenium-webdriver": "^4.36.0"
   },
   "dependencies": {
-    "wait-on": "^8.0.3"
+    "wait-on": "^9.0.1"
   }
 }
