@@ -3,7 +3,7 @@
 import 'femtocrank/style.css'
 import './style.css'
 
-import 'iconify-icon';
+import 'iconify-icon'
 
 import { createClient } from '@connectrpc/connect'
 import { createConnectTransport } from '@connectrpc/connect-web'
@@ -39,7 +39,8 @@ function setupVue () {
 function main () {
   initClient()
 
-  checkWebsocketConnection()
+  // Expose websocket connection function globally so App.vue can call it after successful init
+  window.checkWebsocketConnection = checkWebsocketConnection
 
   setupVue()
 
