@@ -1,4 +1,4 @@
-// @generated by protoc-gen-es v2.7.0
+// @generated by protoc-gen-es v2.9.0
 // @generated from file olivetin/api/v1/olivetin.proto (package olivetin.api.v1, syntax proto3)
 /* eslint-disable */
 
@@ -1306,6 +1306,16 @@ export declare type InitResponse = Message<"olivetin.api.v1.InitResponse"> & {
    * @generated from field: string banner_css = 20;
    */
   bannerCss: string;
+
+  /**
+   * @generated from field: bool show_diagnostics = 21;
+   */
+  showDiagnostics: boolean;
+
+  /**
+   * @generated from field: bool show_log_list = 22;
+   */
+  showLogList: boolean;
 };
 
 /**
