@@ -25,18 +25,13 @@ describe('config: empty dashboards are hidden', function () {
   it('Test hidden dashboard', async function () {
     await getRootAndWait()
 
-    const title = await webdriver.getTitle()
-    expect(title).to.be.equal("OliveTin")
-
     await openSidebar()
+
+    const title = await webdriver.getTitle()
+    expect(title).to.be.equal("Actions - OliveTin")
 
     const navigationLinks = await getNavigationLinks()
     expect(navigationLinks).to.not.be.empty
-    expect(navigationLinks.length).to.be.equal(1, 'Expected the nav to only have 1 link')
-    
-    const firstLinkId = await navigationLinks[0].getAttribute('id')
-
-    expect(firstLinkId).to.be.equal('showActions', 'Expected the first link to be the actions link')
-
+    expect(navigationLinks.length).to.be.equal(4, 'Expected the nav to only have 4 links')
   })
 })
