@@ -6,6 +6,7 @@ import {
   getRootAndWait,
   getActionButtons,
   takeScreenshotOnFailure,
+  openSidebar,
 } from '../lib/elements.js'
 
 describe('config: general', function () {
@@ -25,25 +26,18 @@ describe('config: general', function () {
     await webdriver.get(runner.baseUrl())
 
     const title = await webdriver.getTitle()
-    expect(title).to.be.equal("OliveTin")
-  })
-
-  it('Page title2', async function () {
-    /*
-    await webdriver.get(runner.baseUrl())
-
-    const title = await webdriver.getTitle()
-    expect(title).to.be.equal("OliveTin")
-    */
+    expect(title).to.be.equal("Actions - OliveTin")
   })
 
   it('navbar contains default policy links', async function () {
     await getRootAndWait()
+    await openSidebar()
 
-    const logListLink = await webdriver.findElements(By.css('[href="/logs"]'))
-    expect(logListLink).to.not.be.empty
 
-    const diagnosticsLink = await webdriver.findElements(By.css('[href="/diagnostics"]'))
+    const logsLink = await webdriver.findElements(By.css('a[href="/logs"]'))
+    const diagnosticsLink = await webdriver.findElements(By.css('a[href="/diagnostics"]'))
+
+    expect(logsLink).to.not.be.empty
     expect(diagnosticsLink).to.not.be.empty
   })
 
@@ -56,13 +50,22 @@ describe('config: general', function () {
   it('Default buttons are rendered', async function() {
     await getRootAndWait()
 
-    const buttons = await getActionButtons()
+    await webdriver.wait(new Condition('wait for action buttons', async () => {
+      const btns = await webdriver.findElements(By.css('[title="dir-popup"], [title="cd-passive"], .action-button button'))
+      return btns.length >= 1
+    }), 10000)
 
-    expect(buttons).to.have.length(8)
+    const buttons = await getActionButtons()
+    expect(buttons.length).to.be.greaterThanOrEqual(4)
   })
 
   it('Start dir action (popup)', async function () {
     await getRootAndWait()
+
+    await webdriver.wait(new Condition('wait for dir-popup button', async () => {
+      const btns = await webdriver.findElements(By.css('[title="dir-popup"]'))
+      return btns.length === 1
+    }), 10000)
 
     const buttons = await webdriver.findElements(By.css('[title="dir-popup"]'))
 
@@ -74,19 +77,20 @@ describe('config: general', function () {
 
     buttonCMD.click()
 
-    const dialog = await webdriver.findElement(By.id('execution-results-popup'))
-    expect(await dialog.isDisplayed()).to.be.true
-
-    const title = await webdriver.findElement(By.id('execution-dialog-title'))
-    expect(await webdriver.wait(until.elementTextIs(title, 'dir-popup'), 2000))
-
-    const dialogErr = await webdriver.findElement(By.id('big-error'))
-    expect(dialogErr).to.not.be.null
-    expect(await dialogErr.isDisplayed()).to.be.false
+    // New UI navigates to /logs/<id> instead of showing old dialog
+    await webdriver.wait(new Condition('wait navigate to logs', async () => {
+      const url = await webdriver.getCurrentUrl()
+      return url.includes('/logs/')
+    }), 8000)
   })
 
   it('Start cd action (passive)', async function () {
     await getRootAndWait()
+
+    await webdriver.wait(new Condition('wait for cd-passive button', async () => {
+      const btns = await webdriver.findElements(By.css('[title="cd-passive"]'))
+      return btns.length === 1
+    }), 10000)
 
     const buttons = await webdriver.findElements(By.css('[title="cd-passive"]'))
 
@@ -98,16 +102,10 @@ describe('config: general', function () {
 
     buttonCMD.click()
 
-    const dialog = await webdriver.findElement(By.id('execution-results-popup'))
-    expect(await dialog.isDisplayed()).to.be.false
-
-    const title = await webdriver.findElement(By.id('execution-dialog-title'))
-    expect(await title.getAttribute('innerText')).to.be.equal('?')
-
-    const dialogErr = await webdriver.findElement(By.id('big-error'))
-	console.log("big error is: " + dialogErr.innerHTML)
-    expect(dialogErr).to.not.be.null
-    expect(await dialogErr.isDisplayed()).to.be.false
+    // Should not navigate to logs for passive action
+    await webdriver.sleep(500)
+    const url = await webdriver.getCurrentUrl()
+    expect(url.includes('/logs/')).to.be.false
   })
 
 })
