@@ -1,6 +1,6 @@
 import { describe, it, before, after } from 'mocha'
 import { expect } from 'chai'
-import { By, until } from 'selenium-webdriver'
+import { By, until, Condition } from 'selenium-webdriver'
 import { 
   getRootAndWait, 
   getActionButtons,
@@ -24,6 +24,12 @@ describe('config: multipleDropdowns', function () {
   it('Multiple dropdowns are possible', async function() {
     await getRootAndWait()
 
+    // Wait for action buttons to be rendered
+    await webdriver.wait(new Condition('wait for action buttons', async () => {
+      const btns = await webdriver.findElements(By.css('.action-button button'))
+      return btns.length >= 2
+    }), 10000)
+
     const buttons = await getActionButtons()
 
     let button = null
@@ -40,11 +46,20 @@ describe('config: multipleDropdowns', function () {
 
     await button.click()
 
-    const dialog = await webdriver.findElement(By.id('argument-popup'))
+    // Wait for navigation to argument form page
+    await webdriver.wait(new Condition('wait for argument form page', async () => {
+      const url = await webdriver.getCurrentUrl()
+      return url.includes('/actionBinding/') && url.includes('/argumentForm')
+    }), 8000)
 
-    await webdriver.wait(until.elementIsVisible(dialog), 3500)
+    // Wait for form elements to be rendered
+    await webdriver.wait(new Condition('wait for form elements', async () => {
+      const selects = await webdriver.findElements(By.tagName('select'))
+      return selects.length >= 2
+    }), 5000)
 
-    const selects = await dialog.findElements(By.tagName('select'))
+    // Find the select elements after the wait condition
+    const selects = await webdriver.findElements(By.tagName('select'))
 
     expect(selects).to.have.length(2)
     expect(await selects[0].findElements(By.tagName('option'))).to.have.length(2)
