@@ -1,5 +1,5 @@
 import { describe, it, before, after } from 'mocha'
-import { assert } from 'chai'
+import { assert, expect } from 'chai'
 import { By } from 'selenium-webdriver'
 import {
   getRootAndWait,
@@ -29,22 +29,22 @@ describe('config: onlyDashboards', function () {
     await openSidebar()
 
     const navLinks = await getNavigationLinks()
+    expect(navLinks).to.not.be.empty
 
-    const actionsLink = navLinks[0];
-    assert.isNotNull(actionsLink, 'Actions link should not be null')
-    assert.equal(await actionsLink.getAttribute('title'), 'Actions', 'Actions link should have the title "Actions"')
-    assert.isFalse(await actionsLink.isDisplayed(), 'Actions link should not be displayed when there are only dashboards')
+    for (const link of navLinks) {
+      console.log(await link.getAttribute('title'))
+    }
+
+    const firstLink = await navLinks[0];
+    assert.isNotNull(firstLink, 'Actions link should not be null')
+
+    assert.equal(await firstLink.getAttribute('title'), 'My Dashboard', 'First link should have the title "My Dashboard"')
 
     const firstDashboardLink = await webdriver.findElement(By.css('li[title="My Dashboard"]'), 'The first dashboard link should be present')
     assert.isNotNull(firstDashboardLink, 'First dashboard link should not be null')
     assert.isTrue(await firstDashboardLink.isDisplayed(), 'First dashboard link should be displayed')
     
-    const actionButtons = await getActionButtons()
-
-    assert.isArray(actionButtons, 'Action buttons should be an array')
-    assert.lengthOf(actionButtons, 0, 'Action buttons should be empty when everything is added to the dashboard')
-
-    const actionButtonsOnDashboard = await getActionButtons('MyDashboard')
+    const actionButtonsOnDashboard = await getActionButtons()
     assert.isArray(actionButtonsOnDashboard, 'Action buttons on dashboard should be an array')
     assert.lengthOf(actionButtonsOnDashboard, 3, 'Action buttons on dashboard should have 3 buttons')
   })
