@@ -10,7 +10,6 @@ let metrics = [
   {'name': 'olivetin_actions_requested_count', 'type': 'counter', 'desc': 'The actions requested count'},
   {'name': 'olivetin_config_action_count', 'type': 'gauge', 'desc': 'The number of actions in the config file'},
   {'name': 'olivetin_config_reloaded_count', 'type': 'counter', 'desc': 'The number of times the config has been reloaded'},
-  {'name': 'olivetin_sv_count', 'type': 'gauge', 'desc': 'The number entries in the sv map'},
 ]
 
 describe('config: prometheus', function () {
@@ -27,7 +26,7 @@ describe('config: prometheus', function () {
   });
 
   it('Metrics are available with correct types', async () => {
-    webdriver.get(runner.metricsUrl())
+    await webdriver.get(runner.metricsUrl())
     const prometheusOutput = await webdriver.findElement(By.tagName('pre')).getText()
 
     expect(prometheusOutput).to.not.be.null
