@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.8
+// 	protoc-gen-go v1.36.9
 // 	protoc        (unknown)
 // source: olivetin/api/v1/olivetin.proto
 
@@ -3001,6 +3001,8 @@ type InitResponse struct {
 	EffectivePolicy           *EffectivePolicy       `protobuf:"bytes,18,opt,name=effective_policy,json=effectivePolicy,proto3" json:"effective_policy,omitempty"`
 	BannerMessage             string                 `protobuf:"bytes,19,opt,name=banner_message,json=bannerMessage,proto3" json:"banner_message,omitempty"`
 	BannerCss                 string                 `protobuf:"bytes,20,opt,name=banner_css,json=bannerCss,proto3" json:"banner_css,omitempty"`
+	ShowDiagnostics           bool                   `protobuf:"varint,21,opt,name=show_diagnostics,json=showDiagnostics,proto3" json:"show_diagnostics,omitempty"`
+	ShowLogList               bool                   `protobuf:"varint,22,opt,name=show_log_list,json=showLogList,proto3" json:"show_log_list,omitempty"`
 	unknownFields             protoimpl.UnknownFields
 	sizeCache                 protoimpl.SizeCache
 }
@@ -3173,6 +3175,20 @@ func (x *InitResponse) GetBannerCss() string {
 		return x.BannerCss
 	}
 	return ""
+}
+
+func (x *InitResponse) GetShowDiagnostics() bool {
+	if x != nil {
+		return x.ShowDiagnostics
+	}
+	return false
+}
+
+func (x *InitResponse) GetShowLogList() bool {
+	if x != nil {
+		return x.ShowLogList
+	}
+	return false
 }
 
 type AdditionalLink struct {
@@ -3799,7 +3815,7 @@ const file_olivetin_api_v1_olivetin_proto_rawDesc = "" +
 	"\x16GetDiagnosticsResponse\x12 \n" +
 	"\vSshFoundKey\x18\x01 \x01(\tR\vSshFoundKey\x12&\n" +
 	"\x0eSshFoundConfig\x18\x02 \x01(\tR\x0eSshFoundConfig\"\r\n" +
-	"\vInitRequest\"\xac\a\n" +
+	"\vInitRequest\"\xfb\a\n" +
 	"\fInitResponse\x12\x1e\n" +
 	"\n" +
 	"showFooter\x18\x01 \x01(\bR\n" +
@@ -3824,7 +3840,9 @@ const file_olivetin_api_v1_olivetin_proto_rawDesc = "" +
 	"\x10effective_policy\x18\x12 \x01(\v2 .olivetin.api.v1.EffectivePolicyR\x0feffectivePolicy\x12%\n" +
 	"\x0ebanner_message\x18\x13 \x01(\tR\rbannerMessage\x12\x1d\n" +
 	"\n" +
-	"banner_css\x18\x14 \x01(\tR\tbannerCss\"8\n" +
+	"banner_css\x18\x14 \x01(\tR\tbannerCss\x12)\n" +
+	"\x10show_diagnostics\x18\x15 \x01(\bR\x0fshowDiagnostics\x12\"\n" +
+	"\rshow_log_list\x18\x16 \x01(\bR\vshowLogList\"8\n" +
 	"\x0eAdditionalLink\x12\x14\n" +
 	"\x05title\x18\x01 \x01(\tR\x05title\x12\x10\n" +
 	"\x03url\x18\x02 \x01(\tR\x03url\"L\n" +
